+++ RELEASENOTES.md
@@ -1,5 +1,6 @@
 ## 0.11.1
 
+* don't run install if mulle-clang is already installed
 * adapt to mulle-configuration 3.1